docflow/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gigachat>=0.1.15
pydantic>=2.5.0
aiofiles>=23.2.0
aiosqlite>=0.19.0
pandas>=2.1.0
openpyxl>=3.1.2
dicttoxml>=1.7.16
python-multipart>=0.0.9
python-dotenv>=1.0.0
httpx>=0.26.0
